# apps/express-modular-architecture/src/app/modules/swagger/output/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.1",
    "title": "Penguin API",
    "description": "Documentation automatically generated by the <b>penguin-swagger-autogen</b> module."
  },
  "host": "127.0.0.1:2000",
  "basePath": "/api",
  "tags": [
    {
      "name": "User",
      "description": "Endpoints"
    }
  ],
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/v1/tests/": {
      "get": {
        "tags": ["__template"],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/auth/test": {
      "get": {
        "tags": ["Authentication"],
        "description": "XXXXXXXXXXXX",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/auth/hello2": {
      "get": {
        "tags": ["Authentication"],
        "description": "XXXXXXXXXXXX",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "description": "Sign in a specific user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "description": "Endpoint to sign up a specific user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "confirmPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/auth/profile": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/auth/users": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/categories/hello": {
      "get": {
        "tags": ["User"],
        "description": "XXXXXXXXXXXX",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/media/": {
      "get": {
        "tags": ["media"],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": ["media"],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/media/files/{file}": {
      "get": {
        "tags": ["media"],
        "description": "",
        "parameters": [
          {
            "name": "file",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/v1/media/upload": {
      "post": {
        "tags": ["media"],
        "description": "",
        "produces": [
          "application/json\"charset=utf-8fileSizeNofileidwWrongstartbyteadatacloseUploadfinisheduploadedFileunfinished,stoppedatFileunfinished,stoppedaterrorfileStreamerrorFileerrormedia"
        ],
        "parameters": [
          {
            "name": "x-file-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-start-byte",
            "in": "header",
            "type": "string"
          },
          {
            "name": "name",
            "in": "header",
            "type": "string"
          },
          {
            "name": "size",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/v1/media/convert": {
      "post": {
        "tags": ["media"],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v1/media/status": {
      "get": {
        "tags": ["media"],
        "description": "",
        "parameters": [
          {
            "name": "x-file-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "name",
            "in": "header",
            "type": "string"
          },
          {
            "name": "size",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/news/": {
      "get": {
        "tags": ["news"],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": ["news"],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "published": {
                  "example": "any"
                },
                "draft": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/news/update/{id}": {
      "put": {
        "tags": ["news"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/news/show": {
      "get": {
        "tags": ["news"],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "type": "string"
          },
          {
            "name": "_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "slug",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/news/update": {
      "put": {
        "tags": ["news"],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "type": "string"
          },
          {
            "name": "_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "slug",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/news/delete": {
      "delete": {
        "tags": ["news"],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "type": "string"
          },
          {
            "name": "_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "slug",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/tasks/": {
      "get": {
        "tags": ["tasks"],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": ["tasks"],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/tasks/update/{id}": {
      "put": {
        "tags": ["tasks"],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/tasks/show": {
      "get": {
        "tags": ["tasks"],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "type": "string"
          },
          {
            "name": "_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "slug",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/tasks/update": {
      "put": {
        "tags": ["tasks"],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "type": "string"
          },
          {
            "name": "_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "slug",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/tasks/delete": {
      "delete": {
        "tags": ["tasks"],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "type": "string"
          },
          {
            "name": "_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "slug",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {}
}